Add function to search a list of dicts
Added a function called search_lod_for_key_val that returns a generator object
containing the dictionaries that contain the given key/value pair. It is kind
of a pointless function because the equivalent generator-creating expression
is fewer characters in length than the function call, but the function
signature is maybe a little more readable.